+++ SendBirdUIKit.framework/Headers/SBUInviteUserViewController.swift
@@ -254,6 +254,12 @@ open class SBUInviteUserViewController: UIViewController {
         self.updateStyles()
     }
     
+    open override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        
+        SBULoading.stop()
+    }
+    
     deinit {
         SBULog.info("")
     }